LayoutTests/fast/dom/location-valueOf-after-object-prototype-update.html

<!DOCTYPE html>
<html>
<body>
<script src="../../resources/js-test-pre.js"></script>
<script>
description("Tests the value of location.valueOf after Object.prototype.valueOf has been deleted.");

var objectProtoValueOf = Object.prototype.valueOf;
delete Object.prototype.valueOf;
shouldBe("location.valueOf", "objectProtoValueOf");
</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
